% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{beamer_USCImage}
\alias{beamer_USCImage}
\title{Convert to a Beamer presentation using USC IMAGE theme}
\usage{
beamer_USCImage(background = NULL, ...)
}
\arguments{
\item{background}{Filepath to a background image (see details).}

\item{...}{Further parameters to be passed to \code{beamer_document}}
}
\value{
Whatever \code{beamer_presentation} returns.
}
\description{
Essentially, a wrapper of \code{\link[rmarkdown:beamer_presentation]{beamer_presentation}}
in the \CRANpkg{rmarkdown} package.
}
\details{
The current defaul list of arguments passed via \code{...} is:

\tabular{ll}{
\code{theme}      \tab \code{"default"} \cr
\code{colortheme} \tab \code{"uscimage"} \cr
\code{toc}        \tab \code{TRUE} \cr
\code{highlight}  \tab \code{"tango"} \cr
\code{template}   \tab \Sexpr[results=text]{system.file("templates/uscimage_beamer.tex", package = "uscimage")} \cr
\code{slide_level}\tab \code{2} \cr
\code{includes}   \tab \code{rmarkdown::includes()}
}

If \code{colortheme = "uscimage"}, the argument is set to \code{"default"} and then the
file \Sexpr{system.file("templates/beamercolorthemeuscimage.sty", package = "uscimage")} is
passed to \code{includes$in_header} so that the USCImage color theme overwrites the default
theme.

The default \code{background} image passed is located in
\Sexpr{system.file("templates/uscimage_pptx_background.jpg", package = "uscimage")}.
}
\author{
George G. Vega Yon
}
