% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{beamer_USCImage}
\alias{beamer_USCImage}
\title{Convert to a Beamer presentation using USC IMAGE theme}
\usage{
beamer_USCImage(background = NULL, ...)
}
\arguments{
\item{background}{Filepath to a background image (see details).}

\item{...}{Further parameters to be passed to \code{beamer_document}}
}
\description{
Convert to a Beamer presentation using USC IMAGE theme
}
\details{
By default the template used is in
\code{system.file("templates/uscimage_beamer.tex")} and the
background used is in
\code{system.file("templates/uscimage_pptx_background.jpg")}
}
\author{
George G. Vega Yon
}
